Refactor services and enhance security checks
Refactored user and account management services, adding permission checks for `SuperAdmin` in various methods. Updated constants and method names to reflect these changes. Improved async handling in multiple Razor components and added new indicators for role types. Introduced a new `AccountService` class and interface for user-related operations.

+++ src/Backend/FluentCMS.Services/ApiTokenService.cs
@@ -13,21 +13,30 @@ public interface IApiTokenService : IAutoRegisterService
     Task<ApiToken> Validate(string apiKey, string apiSecret, CancellationToken cancellationToken = default);
 }
 
-public class ApiTokenService(IApiTokenRepository apiTokenRepository, IApiTokenProvider apiTokenProvider, IMessagePublisher messagePublisher) : IApiTokenService
+public class ApiTokenService(IApiTokenRepository apiTokenRepository, IApiTokenProvider apiTokenProvider, IMessagePublisher messagePublisher, IPermissionManager permissionManager) : IApiTokenService
 {
     public async Task<IEnumerable<ApiToken>> GetAll(CancellationToken cancellationToken = default)
     {
+        if (!await permissionManager.HasAccess(GlobalPermissionAction.SuperAdmin, cancellationToken))
+            throw new AppException(ExceptionCodes.PermissionDenied);
+
         return await apiTokenRepository.GetAll(cancellationToken);
     }
 
     public async Task<ApiToken> GetById(Guid tokenId, CancellationToken cancellationToken = default)
     {
+        if (!await permissionManager.HasAccess(GlobalPermissionAction.SuperAdmin, cancellationToken))
+            throw new AppException(ExceptionCodes.PermissionDenied);
+
         return await apiTokenRepository.GetById(tokenId, cancellationToken) ??
             throw new AppException(ExceptionCodes.ApiTokenNotFound);
     }
 
     public async Task<ApiToken> Create(ApiToken apiToken, CancellationToken cancellationToken = default)
     {
+        if (!await permissionManager.HasAccess(GlobalPermissionAction.SuperAdmin, cancellationToken))
+            throw new AppException(ExceptionCodes.PermissionDenied);
+
         var existingApiToken = await apiTokenRepository.GetByName(apiToken.Name, cancellationToken);
         if (existingApiToken != null)
             throw new AppException(ExceptionCodes.ApiTokenNameIsDuplicated);
@@ -45,6 +54,9 @@ public class ApiTokenService(IApiTokenRepository apiTokenRepository, IApiTokenPr
 
     public async Task<ApiToken> Update(ApiToken apiToken, CancellationToken cancellationToken = default)
     {
+        if (!await permissionManager.HasAccess(GlobalPermissionAction.SuperAdmin, cancellationToken))
+            throw new AppException(ExceptionCodes.PermissionDenied);
+
         var existingApiToken = await apiTokenRepository.GetById(apiToken.Id, cancellationToken) ??
             throw new AppException(ExceptionCodes.ApiTokenNotFound);
 
@@ -66,6 +78,9 @@ public class ApiTokenService(IApiTokenRepository apiTokenRepository, IApiTokenPr
 
     public async Task<ApiToken> Delete(Guid tokenId, CancellationToken cancellationToken = default)
     {
+        if (!await permissionManager.HasAccess(GlobalPermissionAction.SuperAdmin, cancellationToken))
+            throw new AppException(ExceptionCodes.PermissionDenied);
+
         var deleted = await apiTokenRepository.Delete(tokenId, cancellationToken) ??
             throw new AppException(ExceptionCodes.ApiTokenUnableToDelete);
 
@@ -76,6 +91,8 @@ public class ApiTokenService(IApiTokenRepository apiTokenRepository, IApiTokenPr
 
     public async Task<ApiToken> Validate(string apiKey, string apiSecret, CancellationToken cancellationToken = default)
     {
+        // there is not need to check permissions here
+        // as this method is used for API authentication
         var apiToken = await apiTokenRepository.GetByKey(apiKey, cancellationToken) ??
             throw new AppException(ExceptionCodes.ApiTokenNotFound);
 
@@ -96,6 +113,9 @@ public class ApiTokenService(IApiTokenRepository apiTokenRepository, IApiTokenPr
 
     public async Task<ApiToken> RegenerateSecret(Guid id, CancellationToken cancellationToken = default)
     {
+        if (!await permissionManager.HasAccess(GlobalPermissionAction.SuperAdmin, cancellationToken))
+            throw new AppException(ExceptionCodes.PermissionDenied);
+
         var apiToken = await apiTokenRepository.GetById(id, cancellationToken) ??
             throw new AppException(ExceptionCodes.ApiTokenNotFound);
